Codehover add row

Codehover add row to the table and images to be linked to each row. You can create multiple lines piping various ch_row.

ch_row(.data = "", text = "", img = "", url = FALSE, ...)

Arguments

.data
Codehover object (text string) created by ch_int.

text
(string) Text to be display inside the table row.

img
(string) Path to the image. If url = TRUE the img should be a url link to a image.

url
(Boolean) Default FALSE, if TRUE indicates that the image is hosted somewhere else and it should not be embeded in the HTML page via base64.

...
Adicional paramethers

Value

The populated HTML table to be passed to another ch_row function or to the ch_out function to finnish the table.